pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "linkedin-content-engine"
version = "2.0.0"
requires-python = ">=3.11"
authors = [{ name = "Sarathi" }]
dependencies = [
    "python-dotenv",
    "pydantic",
    "langgraph",
    "langchain",
    "langchain-community",
    "langchain-mcp-adapters",
    "langchain-openai",
    "google-adk",
    "langfuse",
    "openinference-instrumentation-google-adk",
]

[tool.setuptools.packages.find]
include = ["research_agent*", "content_agent*", "pipeline*", "shared*"]
